Added Circle support to turtle-svg. Added overwrite prompt to animate.sh. Added Video link to readme.

+++ animate.sh
@@ -1,4 +1,24 @@
-#!/bin/sh
+#!/bin/bash
+
+#---------------
+# Dependencies
+#---------------
+
+dependencies="bc ffmpeg convert cargo"
+
+for dep in $dependencies; do
+    echo "Checking whether '$dep' is installed"
+    if ! type $dep > /dev/null; then
+        echo "'$dep' not found, it is required by this script, exiting!"
+        exit 1
+    fi
+done
+
+echo "All dependencies found, Congratulations!"
+
+#---------------
+# Builds
+#---------------
 
 echo "Testing whether you have release versions of turtle and spiral compiled"
 
@@ -8,49 +28,44 @@ echo "$DIR"
 already_built=true
 
 if [ ! -d $DIR/target/release ]; then
-    echo "Target directory not found."
-    mkdir -p target/release
+    echo "Target directory not found, creating it now"
+    mkdir -p target/releaseA
 fi
 
 if [ ! -f $DIR/target/release/spiral ]; then
     echo "Spiral binary not found"
     already_built=false
-fi
-
-if [ ! -f $DIR/target/release/turtle-svg ]; then
+elif [ ! -f $DIR/target/release/turtle-svg ]; then
     echo "Turtle binary not found"
     already_built=false
 fi
 
 if ! $already_built; then
-    echo "Not already built, attempting build."
-    if ! type cargo > /dev/null; then
-        echo "You don't have cargo installed, exiting!"
-        exit 1
-    else 
-        if cargo build --release; then
-            echo "Build succesful"
-        fi
+    echo "Attempting to compile turtle-svg and spiral"
+    if cargo build --release; then
+        echo "Build succesful"
     fi
 else 
-    echo "Already built."
+    echo "Already built"
 fi
 
 if [ -f $DIR/out.mp4 ]; then
-    rm $DIR/out.mp4
+    read -p "out.mp4 already exists, overwrite it? (y/n) " -n 1 -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        rm $DIR/out.mp4
+    else
+        exit 0
+    fi
 fi
 
-if ! type bc > /dev/null; then
-    echo "Please install 'bc', exiting"
-    exit 1
-fi 
-
 # This is the logic that makes the spiral changes slow down (to 0)
-# when the spiral reaches 90 degree angles this function produces 
-# numbers between 0 and 180 it's a sinusoidal generator, with 90 
-# being at the peak of the wave, where the direction of speed-up is 
-# flipped, to produce the same effect but in reverse, for the other
-# 90 degrees.
+# when the spiral reaches 90 degree angles. This function produces 
+# numbers between 0 and 180. It's a sinusoidal generator, with 90 
+# being at the peak of the sine wave. It slows down until it's at
+# 90, then speeds up again.
+
 pi=$(echo "scale=10; 4*a(1)" | bc -l)
 halfpi=$(echo "scale=10; $pi / 2" | bc -l)
 sineish () {
@@ -61,28 +76,46 @@ sineish () {
     echo $pi
 
     for i in `seq 0 $step $halfpi`; do
-        echo `echo "scale=8; s($i) * 90" | bc -l`
+        echo `echo "scale=5; s($i) * 90" | bc -l`
     done
     for i in `seq $halfpi -$step 0`; do
-        echo `echo "scale=8; 180 - (s($i) * 90)" | bc -l`
+        echo `echo "scale=5; 180 - (s($i) * 90)" | bc -l`
     done
 }
 
 echo "How many frames do you want this video to be?"
 read frames
 
+echo "At what frame rate?"
+read framerate
+
+PROGRESSIONS="linear sinosoidal"
+echo "Linear or sinosoidal progression?"
+select progression in $PROGRESSIONS; do
+    if [ "$progression" = "linear" ]; then
+        keyframes=`seq 0 $(echo "scale=5; 180 / $frames" | bc -l) 180`
+        break
+    elif [ "$progression" = "sinosoidal" ]; then
+        keyframes=`sineish $frames`
+        break
+    else
+        clear
+        echo "Invalid option"
+    fi
+done
+
 # This looks terrible, sorry, this is how it works:
-# turtle graphics are printed, which are piped into the interpreter
-# the interpreter spits out svgs
+# Turtle graphics are printed, which are piped into the interpreter
+# The interpreter spits out svgs
 # The svgs are converted to pngs by ImageMagick
-# The pngs are converted encoded into h.264 encoded video using FFmpeg
+# The pngs are encoded into h.264 encoded video using FFmpeg
 
 # All this happens without writing a single frame to disk
-# Everything occurs in memory
+# Everything occurs in memory (or swap, if you need it)
 # Praise be unto Unix pipes and complicated shell scripts
 
-{ for i in `sineish $frames`; do
+{ for i in $keyframes; do
     $DIR/target/release/spiral -g 2 -a $i -i 270 | $DIR/target/release/turtle-svg -w 1000 -h 1000 | convert svg: png:- 
-  done 
-} | ffmpeg -f image2pipe -r 30 -vcodec png -i - -qscale:v 12  -pix_fmt yuv420p -vcodec libx264 out.mp4
+  done
+} | ffmpeg -hwaccel vaapi -f image2pipe -r $framerate -vcodec png -i - -c:v libx264 -pix_fmt yuv420p -preset medium -crf 18 out.mp4
